requirements.txt

Flask==2.0.1
Flask-Cors==3.0.10

gunicorn==20.1.0
h5py==3.7.0
Keras-Preprocessing==1.1.2
nltk==3.7
numpy==1.23.2
scikit-learn
tensorflow-gpu==2.10.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
